clean up locks
