+++ pg_search/src/index/directory/blocking.rs
@@ -31,6 +31,7 @@ use tantivy::{
     error::TantivyError,
 };
 
+use super::lock::BlockingLock;
 use crate::index::reader::segment_component::SegmentComponentReader;
 use crate::index::writer::segment_component::SegmentComponentWriter;
 use crate::postgres::storage::block::{
@@ -43,30 +44,6 @@ use crate::postgres::storage::utils::{BM25BufferCache, BM25Page};
 
 /// Defined by Tantivy in core/mod.rs
 pub static META_FILEPATH: Lazy<&'static Path> = Lazy::new(|| Path::new("meta.json"));
-
-/// Custom lock passed to acquire_lock that uses a buffer as a blocking lock
-#[derive(Debug)]
-pub struct BlockingLock {
-    buffer: pg_sys::Buffer,
-}
-
-impl BlockingLock {
-    pub unsafe fn new(relation_oid: pg_sys::Oid, blockno: pg_sys::BlockNumber) -> Self {
-        let cache = BM25BufferCache::open(relation_oid);
-        let buffer = cache.get_buffer(blockno, Some(pg_sys::BUFFER_LOCK_EXCLUSIVE));
-        Self { buffer }
-    }
-}
-
-impl Drop for BlockingLock {
-    fn drop(&mut self) {
-        unsafe {
-            if pg_sys::IsTransactionState() {
-                pg_sys::UnlockReleaseBuffer(self.buffer)
-            }
-        };
-    }
-}
 
 /// Tantivy Directory trait implementation over block storage
 #[derive(Clone, Debug)]
