+++ pg_search/src/index/directory/mod.rs
@@ -1,2 +1,3 @@
 pub mod blocking;
 pub mod channel;
+pub mod lock;
